Remove the Git::Base::Factory module
